# Make OTel the default for metrics