### Description
Use OTel for metrics by default. One can override the behavior to use OpenCensus instead by setting --enable-otel flag to false.
### Link to the issue in case of a bug fix.
b/387898221

Perf impact:
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
|--------|------------|--------------|------------|--------------|--------------|
| Master |  0.25MiB   | 452.23MiB/s  | 1.16MiB/s  |  63.39MiB/s  |  1.29MiB/s   |
|   PR   |  0.25MiB   | 465.06MiB/s  | 1.33MiB/s  |  65.42MiB/s  |  1.29MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 3757.51MiB/s | 77.25MiB/s | 1176.68MiB/s |  79.37MiB/s  |
|   PR   | 48.828MiB  | 3832.38MiB/s | 80.22MiB/s | 1209.13MiB/s |  81.02MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 3906.58MiB/s | 29.89MiB/s | 443.78MiB/s  |  33.62MiB/s  |
|   PR   | 976.562MiB | 3897.71MiB/s | 33.91MiB/s | 573.08MiB/s  |  32.56MiB/s  |
|        |            |              |            |              |              |

### Testing details
1. Manual - Tested the Prometheus endpoint.
2. Unit tests - NA
3. Integration tests - Covered by existing tests